+++ src/native/libs/System.Native.Browser/native/timer.ts
@@ -1,7 +1,7 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-import { } from "./cross-linked"; // ensure ambient symbols are declared
+import { } from "../../Common/JavaScript/cross-linked"; // ensure ambient symbols are declared
 
 export function SystemJS_ScheduleTimer(shortestDueTimeMs: number): void {
     if (DOTNET.lastScheduledTimerId) {
