--- src/native/libs/System.Runtime.InteropServices.JavaScript.Native/native/cross-linked.ts
@@ -1,5 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-
-import { } from "../../Common/JavaScript/cross-linked";
-
